+++ src/Loader/Preprocessors/Ejs/index.js
@@ -2,7 +2,6 @@ const { readFileSync } = require('fs');
 const path = require('path');
 const { loadModule } = require('../../../Common/FileUtils');
 const { stringifyJSON } = require('../../Utils');
-
 const MarkdownFilter = require('../../PreprocessorFilters/markdown');
 
 // replace the partial file and data to load nested included template via the Webpack loader
@@ -37,7 +36,7 @@ const preprocessor = (loaderContext, options) => {
   const Ejs = loadModule(moduleName);
   const { rootContext } = loaderContext;
 
-  const filterOption = {
+  const markdownFilterOptions = {
     highlight: {
       use: {
         module: 'prismjs',
@@ -48,13 +47,11 @@ const preprocessor = (loaderContext, options) => {
     },
   };
 
-  MarkdownFilter.init(filterOption);
-
   Ejs.fileLoader = (file) => {
     const source = readFileSync(file, 'utf-8');
 
     if (file.toLocaleLowerCase().endsWith('.md')) {
-      return MarkdownFilter.apply(source);
+      return MarkdownFilter.getInstance(markdownFilterOptions).apply(source);
     }
 
     return source;