+++ src/Loader/Preprocessors/Pug/filters/markdown.js
@@ -1,12 +1,16 @@
-const markdown = require('../../../PreprocessorFilters/markdown');
-
 class Filter {
+  static MarkdownFilter = null;
+
   static init(options) {
-    markdown.init(options);
+    if (!this.MarkdownFilter) {
+      // lazy load module only on demand
+      this.MarkdownFilter = require('../../../PreprocessorFilters/markdown/MarkdownFilter');
+    }
+    this.MarkdownFilter.init(options);
   }
 
   static apply(content) {
-    return markdown.apply(content);
+    return this.MarkdownFilter.apply(content);
   }
 }